Rename localstorage keys
This was done so that it doesn't interfere/clash with other projects hosted on my website.

For example, the previous version of this project (hosted on the same domain as this one) also used the same naming system for local storage keys, causing it to clash with those stored by this project.